I defined ip_port to be unsigned precisely so as not to have to do this kind of casting.

+++ core_e/zoolib/ZHTTP_Requests.cpp
@@ -159,7 +159,7 @@ ZRef<ZStreamerR> sRequest(const ZRef<Callable_Connect>& iCallable_Connect,
 			Map theResponseHeader;
 			if (not spRequest(
 				theEP->GetStreamW(), theEP->GetStreamR(),
-				iMethod, theHost + sStringf(":%d",int(uint16(thePort))), thePath, iHeader,
+				iMethod, theHost + sStringf(":%d", thePort), thePath, iHeader,
 				iConnectionClose,
 				&theResponseCode, &theResponseHeader, oRawHeader))
 				{ return null; }
